Update BitfinexBrokerage to use BrokerageMultiWebSocketSubscriptionManager
