+++ Brokerages/Bitfinex/BitfinexBrokerage.Messaging.cs
@@ -93,7 +93,20 @@ namespace QuantConnect.Brokerages.Bitfinex
             : base(WebSocketUrl, websocket, restClient, apiKey, apiSecret, "Bitfinex")
         {
             _job = job;
-            SubscriptionManager = new BitfinexSubscriptionManager(this, WebSocketUrl, _symbolMapper);
+
+            _messageHandler = new BrokerageConcurrentMessageHandler<WebSocketMessage>(OnDataMessage);
+
+            SubscriptionManager = new BrokerageMultiWebSocketSubscriptionManager(
+                WebSocketUrl,
+                MaximumSymbolsPerConnection,
+                0,
+                null,
+                () => new BitfinexWebSocketWrapper(null),
+                Subscribe,
+                Unsubscribe,
+                _messageHandler,
+                _connectionRateLimiter);
+
             _symbolPropertiesDatabase = SymbolPropertiesDatabase.FromDataFolder();
             _algorithm = algorithm;
             _aggregator = aggregator;
@@ -147,7 +160,7 @@ namespace QuantConnect.Brokerages.Bitfinex
         }
 
         /// <summary>
-        /// Should be empty, Bitfinex brokerage manages his public channels including subscribe/unsubscribe/reconnect methods using <see cref="BitfinexSubscriptionManager"/>
+        /// Should be empty, Bitfinex brokerage manages his public channels including subscribe/unsubscribe/reconnect methods using <see cref="BrokerageMultiWebSocketSubscriptionManager"/>
         /// Not used in master
         /// </summary>
         /// <param name="symbols"></param>
@@ -248,8 +261,8 @@ namespace QuantConnect.Brokerages.Bitfinex
                     {
                         case "auth":
                             var auth = token.ToObject<AuthResponseMessage>();
-                            var result = string.Equals(auth.Status, "OK", StringComparison.OrdinalIgnoreCase) ? "succeed" : "failed";
-                            Log.Trace($"BitfinexWebsocketsBrokerage.OnMessage: Subscribing to authenticated channels {result}");
+                            var result = string.Equals(auth.Status, "OK", StringComparison.OrdinalIgnoreCase) ? "successful" : "failed";
+                            Log.Trace($"BitfinexBrokerage.OnMessage: Subscribing to authenticated channels {result}");
                             return;
 
                         case "info":
@@ -258,11 +271,11 @@ namespace QuantConnect.Brokerages.Bitfinex
 
                         case "error":
                             var error = token.ToObject<ErrorMessage>();
-                            Log.Error($"BitfinexWebsocketsBrokerage.OnMessage: {error.Level}: {error.Message}");
+                            Log.Error($"BitfinexBrokerage.OnMessage: {error.Level}: {error.Message}");
                             return;
 
                         default:
-                            Log.Trace($"BitfinexWebsocketsBrokerage.OnMessage: Unexpected message format: {e.Message}");
+                            Log.Error($"BitfinexBrokerage.OnMessage: Unexpected message format: {e.Message}");
                             break;
                     }
                 }
@@ -479,7 +492,7 @@ namespace QuantConnect.Brokerages.Bitfinex
         }
 
         /// <summary>
-        /// Should be empty. <see cref="BitfinexSubscriptionManager"/> manages each <see cref="BitfinexWebSocketWrapper"/> individually
+        /// Should be empty. <see cref="BrokerageMultiWebSocketSubscriptionManager"/> manages each <see cref="BitfinexWebSocketWrapper"/> individually
         /// </summary>
         /// <returns></returns>
         protected override IEnumerable<Symbol> GetSubscribed() => new List<Symbol>();
